made 45 degrees sensors into 20 degree sensors
so it stays in current cell for ir control. Put irs back so they fit on the board and won't stick out (need to double check). replaced middle resistors for more space. replaced motor brackets so they're as inward as possible but still able to fit onto the board with the new space available in the middle back